Add DIVUP definition

+++ lib/ncccl/init.h
@@ -56,7 +56,7 @@ ncclResult_t initChannel(struct ncclComm_t* comm, int channelid) {
   // Communication structures with peers.
   NCCLCHECK(ncclCudaCalloc(&channel->devPeers, comm->nRanks));
   NCCLCHECK(ncclCalloc(&channel->peers, comm->nRanks));
-  for (size_t i=0; i<comm->nRanks; ++i) {
+  for (int i=0; i<comm->nRanks; ++i) {
     channel->peers[i].send.comm = comm;
     channel->peers[i].recv.comm = comm;
   }
